# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Cyrano api swagger",
        "version": "1.0.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/v1/project/{project_id}/loyalties/{loyalty_id}/leaderboard": {
            "get": {
                "description": "Retrieve paginated list of members of loyalty",
                "operationId": "loyalty.leaderboard",
                "parameters": [
                    {
                        "name": "project_id",
                        "in": "path",
                        "description": "ID of the project",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "loyalty_id",
                        "in": "path",
                        "description": "ID of the loyalty program",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "not acceptable"
                    },
                    "401": {
                        "description": "unauthorized"
                    },
                    "404": {
                        "description": "project not found"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
...
        "/api/v1/project/{project_id}/loyalties/{loyalty_id}/leaderboard/member/{uuid}/address": {
            "put": {
                "description": "Save loyalty member address in CRM",
                "operationId": "loyalty.leaderboard.member.address",
                "consumes": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "project_id",
                        "in": "path",
                        "description": "ID of the project",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "loyalty_id",
                        "in": "path",
                        "description": "ID of the loyalty program",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "UUID of a member",
                        "required": true,
                        "type": "string",
                        "format": "uuid"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Loyalty Member Address",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/loyalty_member_address"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "not acceptable"
                    },
                    "401": {
                        "description": "unauthorized"
                    },
                    "404": {
                        "description": "project/loyalty/member not found"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
...
        "/api/v1/projects": {
            "post": {
                "summary": "Create a Project",
                "description": "Create a new Project",
                "operationId": "project.create",
                "consumes": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/project"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "not acceptable"
                    },
                    "401": {
                        "description": "unauthorized"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "loyalty_member_address": {
            "description": "Member Address",
            "required": [
                "address"
            ],
            "properties": {
                "address": {
                    "description": "Address",
                    "type": "string"
                }
            },
            "type": "object"
        },
...
        "project": {
            "description": "Project",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "description": "Project name",
                    "type": "string",
                    "example": "Default Poynt Project"
                }
            },
            "type": "object"
        },
...
    }
}